+++ Changelog.md
@@ -3,6 +3,9 @@ Changelog
 
 ### Version 3.1.0 (2022-09-17)
 * Updated bundles version to 3.1
+* Used a Push const for check push type
+* Added option for enabled mercure service
+* Added option in Config checker
 
 ### Version 3.0.0 (2021-02-10)
 * Created new repository and new branch in to Github